+++ pom.xml
@@ -122,19 +122,19 @@
             </plugin>
 
             <!-- FLYWAY -->
-            <plugin>
-                <groupId>org.flywaydb</groupId>
-                <artifactId>flyway-maven-plugin</artifactId>
-                <version>5.1.1</version>
-                <configuration>
-                    <url>jdbc:postgresql://localhost:5432/reksoft</url>
-                    <user>postgres</user>
-                    <password>postgres</password>
-                    <schemas>
-                        <schema>demo</schema>
-                    </schemas>
-                </configuration>
-            </plugin>
+            <!--<plugin>-->
+                <!--<groupId>org.flywaydb</groupId>-->
+                <!--<artifactId>flyway-maven-plugin</artifactId>-->
+                <!--<version>5.1.1</version>-->
+                <!--<configuration>-->
+                    <!--<url>jdbc:postgresql://localhost:5432/reksoft</url>-->
+                    <!--<user>postgres</user>-->
+                    <!--<password>postgres</password>-->
+                    <!--<schemas>-->
+                        <!--<schema>demo</schema>-->
+                    <!--</schemas>-->
+                <!--</configuration>-->
+            <!--</plugin>-->
         </plugins>
     </build>
