--- src/main/java/ru/reksoft/demo/service/AuthService.java
@@ -1,16 +1,0 @@
-package ru.reksoft.demo.service;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
-
-@Service
-public class AuthService {
-
-    private BCryptPasswordEncoder bCryptPasswordEncoder;
-
-    @Autowired
-    public void setBCryptPasswordEncoder(BCryptPasswordEncoder bCryptPasswordEncoder) {
-        this.bCryptPasswordEncoder = bCryptPasswordEncoder;
-    }
-}
